feat(number-input): apply theme
