+++ src/components/NumberInput/types.ts
@@ -1,0 +1,9 @@
+import type { ChangeEvent, HTMLAttributes } from 'react';
+
+export interface NumberInputProps extends HTMLAttributes<HTMLDivElement> {
+    label?: string;
+    value?: number;
+    isError?: boolean;
+    disabled?: boolean;
+    onChange?: (event: ChangeEvent<HTMLInputElement>) => void;
+}
